Cleaned the Code
I updated everything to the last version with added white noise for 0dB
SNR. I also cleaned the code a little and put all the constant
parameters on top of the file.

The problem is still the Bartlett estimation of PSD or the
Noise_Subtraction itself. The Noise PSD estimation is probably good
enough...